Moved test fields into appropriate tests
