+++ test/animalDetails.graphql.test.ts
@@ -1,13 +1,41 @@
 import { expect } from 'chai';
 import supertest from 'supertest';
 import {v4 as uuidv4} from 'uuid';
-import {animalDetailsFields} from './testFields';
 import createAnimal from './helpers/createAnimalHelper';
 
 require('dotenv').config({ path: './test/.env' });
 
 const url = process.env.TEST_URL || 'http://localhost:8081';
 const request = supertest(url);
+
+// eslint-disable-next-line import/prefer-default-export
+export const animalDetailsFields = `
+    {
+        animalId
+        breed {
+            id
+            abbreviation
+            value
+        }
+        species {
+            id
+            value
+        }
+        gender {
+            id
+            value
+        }
+        color {
+            id
+            value
+        }
+        birthDate
+        weight
+        allergy
+        food
+    }
+`;
+
 let animalId: String;
 
 describe('animalDetails Graphql mutations tests', () => {
