+++ test/animalRegistration.graphql.test.ts
@@ -1,13 +1,22 @@
 import { expect } from 'chai';
 import supertest from 'supertest';
 import { v4 as uuidv4 } from 'uuid';
-import { animalRegistrationFields } from './testFields';
 import createAnimal from './helpers/createAnimalHelper';
 
 require('dotenv').config({ path: './test/.env' });
 
 const url = process.env.TEST_URL || 'http://localhost:8081';
 const request = supertest(url);
+
+// eslint-disable-next-line import/prefer-default-export
+export const animalRegistrationFields = `
+    {
+        registrationNo
+        registrationDate
+        status
+    }
+`;
+
 let animalId: String;
 
 describe('animalRegistration Graphql mutations tests', () => {
