+++ test/breed.graphql.test.ts
@@ -1,12 +1,19 @@
 import { expect } from 'chai';
 import supertest from 'supertest';
 import validate from './validators/breed.interface.validator';
-import { breedFields } from './testFields';
 
 require('dotenv').config({ path: './test/.env' });
 
 const url = process.env.TEST_URL || 'http://localhost:8081';
 const request = supertest(url);
+
+const breedFields = `
+    {
+        id
+        abbreviation
+        value
+    }
+`;
 
 describe('GraphQL breed integration tests', () => {
     it('Returns all breeds translation in "lt" with all fields', (done) => {
