+++ test/organizationTask.graphql.test.ts
@@ -1,10 +1,19 @@
 import { expect } from 'chai';
 import supertest from 'supertest';
 import validate from './validators/organizationTask.interface.validator';
-import { organizationTaskFields } from './testFields';
 
 const url = process.env.TEST_URL || 'http://localhost:8081';
 const request = supertest(url);
+
+const organizationTaskFields = `
+    {
+        id,
+        title,
+        description,
+        organization,
+        isDone
+    }
+`;
 
 describe('OrganizationTasks test', () => {
     it('Returns organization tasks list', (done) => {
