+++ test/translationFields.ts
@@ -1,0 +1,7 @@
+// eslint-disable-next-line import/prefer-default-export
+export const translationFields = `
+    {
+        id
+        value
+    }
+`;
